src/main/java/leetcode/Q234_PalindromeLinkedList.java
package leetcode;

public class Q234_PalindromeLinkedList {
    public boolean isPalindrome(ListNode head) {
        if (head == null || head.next == null) {
            return true;
        }

        ListNode slow = head, fast = head;
        while (fast != null && fast.next != null) {
            slow = slow.next;
            fast = fast.next.next;
        }
        // odd length
        if (fast != null) {
            slow = slow.next;
        }

        slow = reverse(slow);
        fast = head;
        while (slow != null) {
            if (slow.val != fast.val) {
                return  false;
            }
            slow = slow.next;
            fast = fast.next;
        }
        return  true;
    }

    ListNode reverse(ListNode head) {
        ListNode prev = null;
        ListNode node = head;
        ListNode next;
        while (node != null) {
            next = node.next;
            node.next = prev;
            prev = node;
            node = next;
        }
        return prev;
    }
}
